Skip generating LavaMoat policy for MMI (#18416)
The MMI build type is still a work in progress, and should not be
expected to build successfully yet, so we haven't committed a LavaMoat
policy to the repository for it yet. The policy update script has been
updated to skip MMI by default for now; we can include it after that
build has reached a point where we are confident it will build
successfully.