Revert "Updated copy changes for display NFT Media (#20380)" (#21109)
This reverts commit 6c1b71b5845bb4fc07d383c1da7fb823e4c8533f.

## **Description**

This is a temporary fix to
https://github.com/MetaMask/MetaMask-planning/issues/1380

The changes needed to justify the copy update in #20380 have not been
made yet. See [this
comment](https://github.com/MetaMask/MetaMask-planning/issues/1380#issuecomment-1739861587)
for more details

We will make future fixes to the nft-controller to justify the copy
changes that are being reverted here. After that, we can revert this
revert.

## **Manual testing steps**

In the "Security and Privacy" tab of settings, we should see a toggle
with `"Enable OpenSea API"` as the label and `"Use OpenSea's API to
fetch NFT data. NFT auto-detection relies on OpenSea's API, and will not
be available when this is turned off."` as the description.

We should be able to search in settings for that toggle.

Switch to french, spanish and chinese languages, and the expected
translations of the above copy should be present.

When the toggle is off, already imported NFTs from opensea should
display no image. Also, when the toggle is off, you should not be able
to click "Refresh List" to detect NFTs.

However, note that if the toggle is off and you manually import NFTs
whose token metadata is stored on a server, a network request is still
made to that server, and if the image of that NFT is stored on IPFS, it
will be displayed. (This is what we will change in the future, before
reverting this revert.

## **Screenshots/Recordings**

_If applicable, add screenshots and/or recordings to visualize the
before and after of your change._

### After
![Screenshot 2023-08-03 at 5 10 30
PM](https://github.com/MetaMask/metamask-extension/assets/39872794/ea629ba1-4169-465f-ab9b-1c282aa8834d)


### Before
![Screenshot 2023-08-03 at 5 09 46
PM](https://github.com/MetaMask/metamask-extension/assets/39872794/c1401ec5-f9e7-4b7b-b10a-66bf20192c80)

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained:
  - [x] What problem this PR is solving.
  - [x] How this problem was solved.
  - [ ] How reviewers can test my changes.
- [x] I’ve indicated what issue this PR is linked to: Fixes #???
- [ ] I’ve included tests if applicable.
- [ ] I’ve documented any added code.
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
- [ ] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.